PHRAS-60_APPLICATIVE-TRASH (#2074)
* if a "_TRASH_" collection exists, "delete" action moves items to this coll.

* removed debug

* PHRAS-60_APPLICATIVE-TRASH
permalinks are disabled when a doc is deleted (moved to _TRASH_).
deleting a record already in _TRASH_ collection really delete it.

+++ lib/Alchemy/Phrasea/Controller/Prod/RecordController.php
@@ -131,6 +131,8 @@ class RecordController extends Controller
         $StoryWZRepository = $this->getStoryWorkZoneRepository();
 
         $deleted = [];
+        /** @var \collection[] $trashCollectionsBySbasId */
+        $trashCollectionsBySbasId = [];
 
         $manager = $this->getEntityManager();
         foreach ($records as $record) {
@@ -148,10 +150,32 @@ class RecordController extends Controller
                     $manager->remove($attachedStory);
                 }
 
+                $sbasId = $record->getDatabox()->get_sbas_id();
+                if(!array_key_exists($sbasId, $trashCollectionsBySbasId)) {
+                    $trashCollectionsBySbasId[$sbasId] = $record->getDatabox()->getTrashCollection();
+                }
                 $deleted[] = $record->getId();
-                $record->delete();
+                if($trashCollectionsBySbasId[$sbasId] !== null) {
+                    if($record->getCollection()->get_coll_id() == $trashCollectionsBySbasId[$sbasId]->get_coll_id()) {
+                        // record is already in trash so delete it
+                        $record->delete();
+                    }
+                    else {
+                        // move to trash collection
+                        $record->move_to_collection($trashCollectionsBySbasId[$sbasId], $this->getApplicationBox());
+                        // disable permalinks
+                        foreach($record->get_subdefs() as $subdef) {
+                            if( ($pl = $subdef->get_permalink()) ) {
+                                $pl->set_is_activated(false);
+                            }
+                        }
+                    }
+                }
+                else {
+                    // no trash collection, delete
+                    $record->delete();
+                }
             } catch (\Exception $e) {
-
             }
         }
